refactor operator UI casework surfaces
