+++ apps/operator-ui/src/taskActions/taskActionCardUtils.ts
@@ -1,0 +1,11 @@
+export function normalizeOptionalString(value: string) {
+  const normalized = value.trim();
+  return normalized.length > 0 ? normalized : undefined;
+}
+
+export function splitIdentifierList(value: string) {
+  return value
+    .split(/[\n,]+/)
+    .map((entry) => entry.trim())
+    .filter(Boolean);
+}
